Renamed ObjectMetadataNode to reflect it is encrypted.
