+++ src/main/java/com/kryptnostic/kodex/v1/client/KryptnosticConnection.java
@@ -5,11 +5,11 @@ import java.security.PublicKey;
 import java.util.UUID;
 
 import com.kryptnostic.kodex.v1.crypto.ciphers.RsaCompressingCryptoService;
-import com.kryptnostic.kodex.v1.crypto.keys.CryptoServiceLoader;
 import com.kryptnostic.kodex.v1.exceptions.types.SecurityConfigurationException;
 import com.kryptnostic.kodex.v1.storage.DataStore;
 import com.kryptnostic.krypto.engine.KryptnosticEngine;
 import com.kryptnostic.storage.v1.http.CryptoKeyStorageApi;
+import com.kryptnostic.v2.crypto.CryptoServiceLoader;
 
 /**
  * The KryptnosticConnection abstracts out persistence, key generation, and server connection information for down
